pyproject.toml
[tool.poetry]
name = "kumapoon"
version = "0.1.0"
description = ""
authors = ["matt <matt76k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.5.1"
pydantic = "^2.1.1"
pyyaml = "^6.0.1"
arcade = "^2.6.17"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"
isort = "^5.12.0"
pyproject-flake8 = "^6.0.0.post1"
mypy = "^1.4.1"
poethepoet = "^0.22.0"

[tool.flake8]
extend-ignore = "E203,E501,"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
py_version = 310
line_length = 120
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[tool.poe.env]
SRC_DIR = "src kumapoon"

[tool.poe.tasks]
lint = "poetry run pflake8 $SRC_DIR"
black = "poetry run black $SRC_DIR"
isort = "poetry run isort $SRC_DIR"
format = ["isort", "black", "lint"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
